% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_div.R
\name{alpha_div}
\alias{alpha_div}
\title{scavenged from microbiome seq to allow tweaking}
\usage{
alpha_div(physeq, method, rarefy.override = 0)
}
\arguments{
\item{physeq}{phyloseq with microbiome data}

\item{method}{string or vector of strings and can be one or more from ("richness", "fisher", "simpson", "invsimpson", shannon", "evenness", "pd")}

\item{rarefy.override}{Richness needs rarefied data. It automatically detects and rarefies to lowest = default. Override with this setting.}
}
\value{
dataframe
}
\description{
scavenged from microbiome seq to allow tweaking
}
\note{
Updates:
2023-07-05 Alx fixed error on abundance table vs df in richness
2022-09-15 Alx fixed check on taxa in rows otherwise it will return rubbish without knowing.
2020-03-02 Alx split bucket into functions
2020-01-01 Alx initial checkin
}
\author{
Alex Bossers \email{a.bossers@uu.nl}
}
