% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alxinit.R
\name{alxinit}
\alias{alxinit}
\title{Initialise environment
Personal preferences to set for instance cowplot theme globally for ggplot2}
\usage{
alxinit()
}
\description{
Initialise environment
Personal preferences to set for instance cowplot theme globally for ggplot2
}
\note{
Updates:
2020-02-03 Alx initial checkin
}
\author{
Alex Bossers \email{a.bossers@uu.nl}
}
