% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_log1p_score.R
\name{cal_log1p_score}
\alias{cal_log1p_score}
\title{Calculate recalibration matrix for heatmaps
Apply transformation function of log avoiding 0 (1p)}
\usage{
cal_log1p_score(x)
}
\arguments{
\item{x}{value}
}
\value{
tranformed value
}
\description{
Calculate recalibration matrix for heatmaps
Apply transformation function of log avoiding 0 (1p)
}
\note{
Updates:
2020-03-02 Alx split bucket into functions
2020-01-01 Alx initial checkin
}
\author{
Alex Bossers \email{a.bossers@uu.nl}
}
