% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_transform.R
\name{clr_transform}
\alias{clr_transform}
\title{Apply the centered log-ratio trasformation on a relative abundance matrix!
Species in columns and samples in rows.
Defined as log^base( x / geometric_mean(rowx) )}
\usage{
clr_transform(x, base = 2, pseudocount = FALSE)
}
\arguments{
\item{x}{OTU/ASV relative abundance matrix species as cols and samples as rows.}

\item{base}{default=base log 2 (can be 10 or exp(1)....}

\item{pseudocount}{default=FALSE applies a pseudocount of MIN(relative abundance)/2 to exact zero relative abundance entries in ASV/OTU table before taking logs.}
}
\value{
transformed matrix
}
\description{
Apply the centered log-ratio trasformation on a relative abundance matrix!
Species in columns and samples in rows.
Defined as log^base( x / geometric_mean(rowx) )
}
\author{
Alex Bossers \email{a.bossers@uu.nl}

joey711 Original idea scavenged from R phyloseq Shiny documentation and slightly adapted https://github.com/joey711/shiny-phyloseq/blob/master/panels/paneldoc/Transform.md

Versions:
20211103 mad pseudocount non default (was not intuitive)
20200909 added default pseudocount to prevent log(0)
20190322 initial scavenge and adaptation'
}
