% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorDistinct.R
\name{colorDistinct}
\alias{colorDistinct}
\title{Generates a distinct scale of n=x elements
based on QUAL Rcolorebrewer set
Manually replaced a few too dominant or too dark colors with other variants}
\usage{
colorDistinct(n, lightfirst = TRUE)
}
\arguments{
\item{n}{number of desired distinct colors}

\item{lightfirst}{boolean. Default TRUE.}
}
\value{
vector of hex colors size N
}
\description{
Generates a distinct scale of n=x elements
based on QUAL Rcolorebrewer set
Manually replaced a few too dominant or too dark colors with other variants
}
\note{
Will modify this function to have repeated HUE variants of other frequently used colorschemes

Updates:
2021-12-02 Alx update two subsequent grey colors to more blueish
2020-03-02 Alx split bucket into functions
2020-01-01 Alx initial checkin
}
\author{
Alex Bossers \email{a.bossers@uu.nl}
}
