% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countTaxrankNAs.R
\name{countTaxrankNAs}
\alias{countTaxrankNAs}
\title{Counts the number of NAs in the taxonomy table per taxonomic rank.
Returns a dataframe having the NA counts and the percentage NAs of total entries per taxrank.
Write optionally the data to a file.}
\usage{
countTaxrankNAs(ps, filename = "")
}
\arguments{
\item{ps}{phyloseq object}

\item{filename}{Optional filename and path if you want to save the output}
}
\value{
data frame
}
\description{
Counts the number of NAs in the taxonomy table per taxonomic rank.
Returns a dataframe having the NA counts and the percentage NAs of total entries per taxrank.
Write optionally the data to a file.
}
\author{
Alex Bossers \email{a.bossers@uu.nl}
}
