% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrare2.R
\name{ggrare2}
\alias{ggrare2}
\title{Rarefaction curve, ggplot style
This will generate a rarefaction plot of the ps object
Custom step-size and incl standard error spread.}
\usage{
ggrare2(
  physeq,
  step = 0,
  label = NULL,
  labelsize = 4,
  color = NULL,
  plot = FALSE,
  parallel = TRUE,
  se = FALSE
)
}
\arguments{
\item{physeq}{phyloseq class object, from which abundance data are extracted}

\item{step}{Step size for sample size in rarefaction curves. Default 1/100 of highest samplesum.}

\item{label}{Default \code{NULL}. Character string. The name of the variable
to map to text labels on the plot. Similar to color option
but for plotting text.}

\item{labelsize}{Default 4.}

\item{color}{(Optional). Default NULL. Character string. The name of the
variable to map to colors in the plot. This can be a sample
variable (among the set returned by
sample_variables(physeq) ) or taxonomic rank (among the set
returned by rank_names(physeq)).
Finally, The color scheme is chosen automatically by
ggplot, but it can be modified afterward with an
additional layer using scale_color_manual.}

\item{plot}{Logical, should the graphic be plotted.}

\item{parallel}{should rarefaction be parallelized (using parallel framework)}

\item{se}{Default FALSE. Logical. Should standard errors be computed and plotted.}
}
\description{
Required initially to have counts as integers. Updated the function to accept rounded num.
This is just a more convenient warpper around ggrare allowing tweaks but also object reload without recalc
}
\details{
Args:
}
\author{
Alex Bossers \email{a.bossers@uu.nl}

gauravsk: The original unmodified ggrare function was scavenged from: https://rdrr.io/github/gauravsk/ranacapa/man/ggrare.html

Updates:
2024-05-30 Alx Add auto step-size and warning at purge samples below stepsize
2021-11-26 Alx Removed the replot option due to often stack errors.
2021-10-18 Alx Integrated the ggrare_replot function into this one.
2018-11-12 Alx Required initially to have counts as integers. Updated the function to accept rounded num.
2018-03-02 Alx Added additional check if sample depth is equal or higher than supplied rarestep to prevent errors
We just prune samples equal or higher and will e EXCLUDED from the plots!
2018-01-01 Alx initial checkin
}
