% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdisp.R
\name{hdisp}
\alias{hdisp}
\title{Multiple tests for homogeneity of dispersions (vegan::betadisper)}
\usage{
hdisp(physeq, distm = "bray", vars = NULL, hellinger = TRUE)
}
\arguments{
\item{physeq}{a \link{phyloseq} object}

\item{distm}{choice of dissimilarity index according to \link{vegdist}. Options are "manhattan", "euclidean", "canberra", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup" , "binomial", "chao", "cao" or "mahalanobis".}

\item{vars}{selected variables from the ps object (should be type factor and 2 or more levels). Default uses all.}

\item{hellinger}{default=TRUE. Performs Hellinger transformation on the otu table.}
}
\value{
function prints a list of p-values and returns a list of \code{betadisper} outputs which can be examined with \code{output[[x]]} to view the output for the xth element of \code{sample_data(physeq)}
}
\description{
Function performs individual, \code{\link{betadisper}}-type tests for homogeneity of dispersions for every element of \code{sample_variables(physeq)}. Community matrix is Hellinger-transformed prior to the calculation of distances using specified matrix.
}
\examples{
q <- hdisp(physeq, distm="bray", vars=c("Type","Date","Health"))
plot(q[[1]])

}
\author{
Alex Bossers \email{a.bossers@uu.nl}

Alx: added return list incl p-values
}
