% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.adonis.R
\name{pairwise.adonis}
\alias{pairwise.adonis}
\title{\link{deprecated} Calculate pairwise PERMANOVAs on phyloseq.}
\usage{
pairwise.adonis(
  x,
  factors,
  sim.function = "vegdist",
  sim.method = "bray",
  p.adjust.m = "bonferroni"
)
}
\arguments{
\item{x}{community table, site by species}

\item{factors}{a column or vector with all factors to be tested pairwise}

\item{sim.function}{function to calculate the similarity matrix. IE. 'vegdist' default is 'vegdist'}

\item{sim.method}{similarity method from daisy or vegdist: default is 'bray'}

\item{p.adjust.m}{p.value correction method, one of the methods supported by p.adjust(); default is 'bonferroni'}
}
\value{
a table with the pairwise factors, F-values, R^2, p.value and adjusted p.value
}
\description{
NOT SURE IF THIS FUNCTION STILL MAKES SENSE. KEPT just in case from SJurburg
}
\note{
DEPRECATED function

scavenged from https://www.researchgate.net/post/How_can_I_do_PerMANOVA_pairwise_contrasts_in_R
}
\examples{
data(iris)
pairwise.adonis(iris[,1:4],iris$Species)

[1] Signif. codes:  Signif. codes:  0 = ***, 0.001 = **, 0.01 = *, 0.05 = .
                   pairs    F.Model        R2 p.value p.adjusted sig
1    setosa vs versicolor  552.51266 0.8493496   0.001      0.003   *
2     setosa vs virginica 1001.54509 0.9108722   0.001      0.003   *
3 versicolor vs virginica   91.82959 0.4837475   0.001      0.003   *

#similarity euclidean from vegdist and holm correction
#pairwise.adonis(x=iris[,1:4],factors=iris$Species,sim.function='vegdist',sim.method='euclidian',p.adjust.m='holm')
adon=pairwise.adonis(physeq_pruned@otu_table@.Data,physeq_pruned@sam_data$Sample_type)

}
\author{
Pedro Martinez Arbizu
}
