% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.permanovas2.R
\name{pairwise.permanovas2}
\alias{pairwise.permanovas2}
\title{Pairwise PERMANOVAs (adonis2) on phyloseq to find which pairs of a more complete set have significantly different community structures}
\usage{
pairwise.permanovas2(
  ps,
  variable,
  sim.function = "vegdist",
  sim.method = "bray",
  p.adjust.m = "BH",
  hellinger = TRUE,
  rngseed = 2202,
  num.cores = "",
  ...
)
}
\arguments{
\item{ps}{phyloseqobject}

\item{variable}{all factors to be tested pairwise "single" or multiple c("A","B")}

\item{sim.function}{function to calculate the similarity matrix (vegdist or daisy). Default is 'vegdist'}

\item{sim.method}{similarity method from daisy or vegdist: default is 'bray'}

\item{p.adjust.m}{p.value correction method, one of the methods supported by p.adjust(); default BH = 'Benjamini Hochberg'}

\item{hellinger}{boolean. Default TRUE to Hellinger transform the "distance" matrix}

\item{rngseed}{The set.seed starting point. Default 2202}

\item{num.cores}{Limit the number of cores to be used. Default = available cores - 1}

\item{...}{Transfer any other common variables to the adonis2 function}
}
\value{
a table with the pairwise factors, F-values, R^2, p.value and adjusted p.values
}
\description{
Pairwise PERMANOVAs (adonis2) on phyloseq to find which pairs of a more complete set have significantly different community structures
}
\note{
Based initially on a function by Pedro Martinez Arbizu

Initially scavenged from https://www.researchgate.net/post/How_can_I_do_PerMANOVA_pairwise_contrasts_in_R
}
\examples{
#library(phyloseq)
#data(GlobalPatterns)
#pairwise.permanovas2( GlobalPatterns, variable=c("SampleType"), 
#                      sim.method = "bray", hellinger = TRUE, p.adjust.m = "BH")
#
## No significant comparisons in this dataset.
#
#> Signif. codes: 0 = ***, 0.001 = **, 0.01 = *, 0.05 = .
#>                                       Pairs   F.Model        R2     Pr(>F)     p.adj sig
#> 1                             Soil vs Feces  7.937235 0.6135186 0.04400000 0.1028571    
#> 2                              Soil vs Skin  5.420047 0.5753737 0.10000000 0.1028571    
#> 3                            Soil vs Tongue  5.832750 0.6603549 0.10000000 0.1028571    
#> 4                        Soil vs Freshwater  4.832812 0.6169958 0.10000000 0.1028571    
#> 5                Soil vs Freshwater (creek)  8.934814 0.6907571 0.10000000 0.1028571    
#> 6                             Soil vs Ocean  8.016178 0.6671155 0.10000000 0.1028571    
#> 7                Soil vs Sediment (estuary)  8.045936 0.6679378 0.10000000 0.1028571    
#> 8                              Soil vs Mock 13.453708 0.7708223 0.10000000 0.1028571    
#> 9                             Feces vs Skin  7.440478 0.5980862 0.02800000 0.1028571    
#>...

}
\author{
Alex Bossers (a.bossers\@uu.nl)
}
