% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_anova.R
\name{perform_anova}
\alias{perform_anova}
\title{Function override in of microbiomeseq to calculate pair-wise anovas allowing more options and Welch-ANOVA}
\usage{
perform_anova(
  df,
  meta_table,
  grouping_column,
  pValueCutoff = 0.05,
  anova.type = "anova"
)
}
\arguments{
\item{df}{dataframe}

\item{meta_table}{the metadata df}

\item{grouping_column}{the actual column to test variables of}

\item{pValueCutoff}{as it says (default 0.05)}

\item{anova.type}{Defaults to classic ANOVA ("anova"). Option to use Welch-ANOVA ("welch").}
}
\value{
list
}
\description{
Function override in of microbiomeseq to calculate pair-wise anovas allowing more options and Welch-ANOVA
}
\note{
For Welch-ANOVA we use the \href{https://www.rdocumentation.org/packages/rstatix}{rstatix package}.

Updates:
2022-08-02 Alx added Welch-ANOVA option
2019-11-01 Alx initial checkin
}
\author{
Alex Bossers \email{a.bossers@uu.nl}

Additional credits go the the original function perform_anova_diversity from \href{https://github.com/umerijaz/microbiomeSeq}{MicrobiomeSeq} by Alfred Ssekagiri which I modified and extended with some convenience options (mainly) for myself.
}
