% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova2.R
\name{permanova2}
\alias{permanova2}
\title{Single (multivariable) PERMANOVA v2 Uses vegan::adonis2. This function is just a convenience wrapper for vegan::adonis2.}
\usage{
permanova2(
  physeq = "",
  distm = "bray",
  formula = "",
  var = "",
  permutations = 999,
  hellinger = TRUE,
  rngseed = 2202,
  ...
)
}
\arguments{
\item{physeq}{phyloseq object. atrix is created using otu_table(physeq); permanovas are computed for each element of sample_variables(physeq)}

\item{distm}{choice of dissimilarity index according to vegan::vegdist. Options are "manhattan", "euclidean", "canberra", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup" , "binomial", "chao", "cao" or "mahalanobis".}

\item{formula}{Defaults to distnmatrix ~ var}

\item{var}{ONE variable. Required if no formula is specified! (should be factor levels >=2).}

\item{permutations}{default 999. Number of permutations to use.}

\item{hellinger}{TRUE. By default performs a Hellinger transformation of the OTU table. Here you can turn it off.}

\item{rngseed}{set.seed number to use for reproducible permutations. Default rngseed=2202.}

\item{...}{Allows additional arguments to the adonis2 function to be taken along. For instance parallel = parallel::detectCores() - 1}
}
\value{
vegan2 default output.
}
\description{
Function performs individual, adonis2 permanova for given variable or (multivariable) formula
}
\note{
if sample_data(physeq) contains NAs, the function will convert these to an additional level "none"
}
\examples{
# data(GlobalPatterns)
# physeq <- GlobalPatterns
# per <- permanova2( physeq, distm="bray", formula="~ Treatment + AnimalNumber", var="",
#                    rngseed="2202", parallel=(parallel::detectCores() - 1) )
# per <- permanova2( physeq, distm="bray", formula="", var="Treatment", rngseed="2202")

}
\author{
Convenience wrapper: Alex Bossers \email{a.bossers@uu.nl}
}
