% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanovas2.R
\name{permanovas2}
\alias{permanovas2}
\title{Multiple Permanovas v2
Uses vegan::adonis2 which returns a slight different object.
V2 now returns a proper data frame of tested vars and R2 and F.}
\usage{
permanovas2(
  physeq,
  distm = "bray",
  perms = 999,
  vars = "",
  hellinger = TRUE,
  rngseed = 2202,
  verbose = TRUE
)
}
\arguments{
\item{physeq}{phyloseq object.}

\item{distm}{choice of dissimilarity index according to vegan::vegdist. Options are "manhattan", "euclidean", "canberra", "bray", "kulczynski", "jaccard",
"gower", "altGower", "morisita", "horn", "mountford", "raup" , "binomial", "chao", "cao" or "mahalanobis".}

\item{perms}{the number of permutations to be performed. Default is 999}

\item{vars}{Empty=default tests all metadata variables (should be factor levels >=2). Otherwise give it a vector c("var1","var2",...etc) of factor variables.}

\item{hellinger}{boolean. Default TRUE to Hellinger transform the "distance" matrix}

\item{rngseed}{set.seed number to use for reproducible permutations. Default rngseed=2202}

\item{verbose}{boolean. Default is TRUE and prints progress.}
}
\value{
a dataframe containing p-values of each comparison in the first column, and the number of levels for that variable in the second.
}
\description{
Function performs individual, adonis-style permanovas for every/selected elements of sample_variables(physeq)
}
\note{
if sample_data(physeq) contains NAs, the function will convert these to an additional level "none"

Added the complete adonis output in verbose mode (alex 20200515)

V2 Moved to adonis2 since adonis will be deprecated soon (20220610)
}
\examples{
# per <- permanovas2(ps, distm="bray")

}
\author{
Alex Bossers \email{a.bossers@uu.nl} based on initial idea by Stephanie Jurburg.
}
