% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylodiv.R
\name{phylodiv}
\alias{phylodiv}
\title{Phylodiv calculates phylogenetic diversity from phyloseq (req tree)}
\usage{
phylodiv(physeq, theta = 0)
}
\arguments{
\item{physeq}{phyloseq class object, from which phylogeny and abundance data are extracted}

\item{theta}{parameter that determines the balance in the Balance Weighted Phylogenetic Diversity (see McCoy and Matsen, 2013)
Theta = 0 corresponds to Faith's PD}
}
\value{
dataframe
}
\description{
Phylodiv calculates phylogenetic diversity from phyloseq (req tree)
}
\note{
Requires a phylogenetic tree in your phyloseq object!
}
\author{
Original author unknown. You? Contact me for credits.

Alex Bossers UU /email{a.bossers@uu.nl}
}
