% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_plot.R
\name{prevalence_plot}
\alias{prevalence_plot}
\title{Plot prevalence plots of certain taxonomic rank from phyloseq object}
\usage{
prevalence_plot(
  ps,
  rank = "Phylum",
  xlog = TRUE,
  ylog = FALSE,
  yrelative = FALSE,
  strbck = "#f0f0f0",
  title = paste0("Taxa (", rank, ") prevalence over all samples (Nsamples=",
    nsamples(ps), ")")
)
}
\arguments{
\item{ps}{a phyloseq object containing cummunity abundance data}

\item{rank}{string, default rank="Phylum"
Must match "Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"}

\item{xlog}{boolean, optional, default xlog=TRUE}

\item{ylog}{boolean, optional, default ylog=FALSE}

\item{yrelative}{boolean, optional, default FALSE. Calculate prevalence as relative or sample number}

\item{strbck}{string, optional, default strbck=#f0f0f0, strip background color override}

\item{title}{string, optional, Default OTU prevalence over all samples (Rank, nsamples=xx)}
}
\value{
ggplot object
}
\description{
Function will return a prevalence plot from each OTU
N occurences in samples (y) and total abundance (x)
Returns a ggplot object
Use the cowplot theme for great results
}
\note{
Prevalence plot based on Callahan et al 2016 (McMurdie and Holmes group); https://f1000research.com/articles/5-1492/v2
}
\examples{
prevplot0 <- prevalence_plot( ps = PhyObject, rank = "Family", xlog = TRUE, ylog = FALSE,
yrelative = FALSE, title = "OTU prev over all samples" )

}
\author{
Alex Bossers \email{a.bossers@uu.nl}

Alex Bossers \email{alex.bossers@wur.nl}
}
