% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_taxa.R
\name{prevalence_taxa}
\alias{prevalence_taxa}
\title{Calculate the prevalence at lowest taxonomic rank (i.e. ASV) from phyloseq. Function is used in the prevalence_plot() function}
\usage{
prevalence_taxa(ps, prev_relative = FALSE)
}
\arguments{
\item{ps}{a phyloseq object containing cummunity abundance data}

\item{prev_relative}{prevalence as numbers of samples or relative to total number of samples}
}
\value{
dataframe object
}
\description{
Function will return a prevalence data frame from each OTU
N occurences in samples (y) and total abundance (x)
Optional the prevalence can be relative (abundance divided by N samples)'
}
\note{
Prevalence plot based on Callahan et al 2016 (McMurdie and Holmes group) https://f1000research.com/articles/5-1492/v2
}
\examples{
prev0 <- prevalence_taxa( ps = PhyObject )

}
\author{
Alex Bossers \email{a.bossers@uu.nl}

Alex Bossers \email{alex.bossers@wur.nl}
'
}
