% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes_error.R
\name{procrustes_error}
\alias{procrustes_error}
\title{Plot Procrustes residuals}
\usage{
procrustes_error(df = "", proc_df, main = "Procrustes residuals")
}
\arguments{
\item{df}{\if{html}{\out{<deprecated>}} do not use. For backward compatibility only.}

\item{proc_df}{the procrustus data frame (from vegan)}

\item{main}{Title to the plot}
}
\value{
Base R plot
}
\description{
Plot Procrustes residuals
}
\note{
Uses the data from procrustes vegan package
}
\author{
Alex Bossers (UU)
}
