% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_plot_bar.R
\name{ps_plot_bar}
\alias{ps_plot_bar}
\title{Convenience wrapper to generate tuned and sorted barplots at various taxonomic levels from phyloseq input}
\usage{
ps_plot_bar(
  ps,
  taxrank = "Phylum",
  top = 10,
  taxglom = TRUE,
  taxglom.next = "",
  sort.stack = TRUE,
  x = "Sample",
  y = "Abundance",
  group = "",
  group_fun = "mean",
  logy = FALSE,
  facet_grid = NULL,
  NArm = FALSE,
  xlab.rel = 0.5,
  legend = TRUE,
  legend.col = 1,
  legend.size = 8,
  title = paste0("Barplot ", taxrank, " - ", if (top > 0) {
     paste0("Top", top)
 }
    else {
     "all taxa"
 }),
  title.center = TRUE
)
}
\arguments{
\item{ps}{phyloseq object}

\item{taxrank}{taxonomic rank to show (default "Phylum")}

\item{top}{top-x to show (default top=10)}

\item{taxglom}{Agglomerate the taxonomic data at the indicated level. Default TRUE}

\item{taxglom.next}{Show subdivision of the bars af a next lower level of taxonomic rank. Default = ""}

\item{sort.stack}{whether or not to sort the actual stacking on overall abundance (default sort.stack=TRUE)}

\item{x}{what to show on x-axis (default x="Sample")}

\item{y}{what to show on y-axis (default y="Abundance")}

\item{group}{metadata grouping variable to "merge_samples" phyloseq function}

\item{group_fun}{default "mean". phyloseq merge_samples will SUM otu_table by default. Here we corrected for the Nsamples per group -> actual MEAN. All other values will result in default behaviour.}

\item{logy}{transform Abundance y-scale to log10 (default FALSE)}

\item{facet_grid}{we want to facet the stacked bars (default facet_grid=NULL, example facet_grid="Kingdom")}

\item{NArm}{whether or not to remove NAs in the data while agglomerating at given taxonomic rank (default NArm=FALSE)}

\item{xlab.rel}{Tweaked size of x-axis labels since readability depends on N samples (default xlab.rel=0.5)}

\item{legend.col}{How many columns the legend may contain (default legend.col=1)}

\item{legend.size}{FOnt size to use in legend. Default 8 pnt.}

\item{title}{Any custom to plot on the figure (default generates "Barplot \if{html}{\out{<taxrank>}} - top\if{html}{\out{<x>}} taxa")}

\item{title.center}{center the title on figure otherwise left adjusted (default=TRUE)}
}
\value{
ggplot figure
}
\description{
Convenience wrapper to generate tuned and sorted barplots at various taxonomic levels from phyloseq input
}
\note{
Updates:
2022-07-01 alx added taxglom options to allow visible subtaxonomic ranks per bar.
2022-05-15 alx added grouping and MEAN grouping options
2021-xx-xx alx many small tweaks. See git history.
2020-03-02 Alx split bucket into functions
2020-01-01 Alx initial checkin

FixMe: sorting abundance of stacks when not using taxglom showing subtaxonomic ranks. Sorting is different.
}
\author{
Alex Bossers \email{a.bossers@uu.nl}
}
