% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy_custom_depth.R
\name{rarefy_custom_depth}
\alias{rarefy_custom_depth}
\title{Convenience function to rarefy a given phyloseq object not to one value for each sample but to allow to rarefy each sample with its own given rarefy target depth.}
\usage{
rarefy_custom_depth(physeq, rarefy_depth, seed = 2202)
}
\arguments{
\item{physeq}{phyloseq object containing at least 2 samples}

\item{rarefy_depth}{vector of counts per sample to target rarefy (match sample order!)}

\item{seed}{which random seed to use}
}
\value{
m.final rarefied phyloseq object
}
\description{
Why? In the exceptional case when you want to proportionally rarefy the data to correct for instance for input volume differences
which may be the case in semi-quantitative measurements of the resistome. Since with resistome for example we do not always want to
work with relative abundances (100\%) and maintaining the ARG total load per sample (FPK or FPKM similar).
}
\note{
version: 1.1
versiondate: 20201112
}
\author{
: a.bossers\@uu.nl
}
