% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sum_plot.R
\name{sample_sum_plot}
\alias{sample_sum_plot}
\title{Make a (coloured) ggplot of sample sums per sample ordered decreasing by sample sum and some stats like 90\% rarefy number on phyloseq}
\usage{
sample_sum_plot(
  ps,
  percentKeep = 90,
  cutoff = FALSE,
  numberKeep = FALSE,
  logy = TRUE,
  color = "",
  size.points = 3,
  stats = TRUE,
  subtitle = "",
  crosshair = TRUE,
  xlab.rel = 0.8,
  namesColumn = ""
)
}
\arguments{
\item{ps}{phyloseq object}

\item{percentKeep}{default 90. Percentage of samples to keep at rarefaction cutoff indication}

\item{cutoff}{default FALSE. Overrides percentKeep and sets the minimum sampleSum to take.}

\item{numberKeep}{default FALSE. Override percentKeep and cutoff and sets the number of samples to set rarefy threshold to}

\item{logy}{default TRUE. Make y-scale log10}

\item{color}{optional metadata variable name to color the points in the graph}

\item{size.points}{default=3. Actual point size for the ggplot.}

\item{stats}{default TRUE. Show median, max and min in top right corner of plot.}

\item{subtitle}{default none. gplot subtitle string}

\item{crosshair}{default TRUE. Show the bluegrey crosshair target on selected cutoff.}

\item{xlab.rel}{Scale the x-axis labels to something readible (default scaling text 0.8x)}

\item{namesColumn}{default SampleNames. Custom metadata column to use as sample names along x-axis.}
}
\value{
ggplot_object
}
\description{
Make a (coloured) ggplot of sample sums per sample ordered decreasing by sample sum and some stats like 90\% rarefy number on phyloseq
}
\author{
\email{a.bossers@uu.nl}
}
