% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveVariousPlots.R
\name{SaveVariousPlots}
\alias{SaveVariousPlots}
\title{Saves ggplot plots with reasonable defaults in png and svg (png has explicit white background (transparency off))}
\usage{
SaveVariousPlots(
  filename_base = "",
  plotobj,
  scale = 1.5,
  width = 40,
  height = 20,
  units = "cm",
  bg_png = "white",
  filetype = ""
)
}
\arguments{
\item{filename_base}{the filename base to save the figures}

\item{plotobj}{ggplot image object}

\item{scale}{default 1.5}

\item{width}{default 25}

\item{height}{default 12}

\item{units}{default cm}

\item{bg_png}{default png background white (alternative "transparent" or a color)}

\item{filetype}{override to custom ggsave filetype (default "")}
}
\value{
ggplot figures saved in png (quick pastes) and svg (vector graphics) or alternative with custom ggplot2 filetype
}
\description{
Figures look great when using cowplot theme as default.
}
\note{
Updates:
2025-02-18 Alx added default white background for PNG only (saved trasparent earlier).
2024-12-21 Alx removed default pdf
2020-03-02 Alx split bucket into functions
2020-01-01 Alx initial checkin
}
\author{
Alex Bossers \email{a.bossers@uu.nl}
}
