% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short.tn.R
\name{short.tn}
\alias{short.tn}
\title{Shorten too long temp OTU names from provided vector.}
\usage{
short.tn(original)
}
\arguments{
\item{original}{The taxon name to shorten as string}
}
\value{
shortened string
}
\description{
Shorten too long temp OTU names from provided vector.
}
\author{
a.bossers@uu.nl

scavenged from / inspired by https://peerj.com/articles/545/#fig-1
}
