% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simper.pretty2.R
\name{simper.pretty2}
\alias{simper.pretty2}
\title{Perform SIMPER and return pretty output using phyloseq (debugged).}
\usage{
simper.pretty2(
  ps,
  interesting,
  perc_cutoff,
  low_cutoff,
  low_val,
  output_name,
  cores = detectCores() - 2
)
}
\arguments{
\item{ps}{phyloseq object having: otu table OTU-names in columns! Function will now transpose.}

\item{interesting}{list the column headers of interest from metrics file ex. c('sloth_sp','type','sp.type')}

\item{perc_cutoff}{\% cutoff desired for SIMPER output, as decimal (i.e. write 50\% as 0.5)}

\item{low_cutoff}{'y' if you want to REMOVE OTUs that contribute to less than 1\% of significance}

\item{low_val}{value of low cutoff (0.01)}

\item{output_name}{name to append to the simper and clean_simper output files}

\item{cores}{number of cores to use. Default limits to cores available minus 2. Windows does not do multi core by default.

Alex: cores can be specified to limit used cores otherwise all available are used (minus 2)!}
}
\description{
Perform SIMPER and return pretty output using phyloseq (debugged).
}
\author{
a.bossers@uu.nl

What it does:
This script is meant to rapidly perform the SIMPER function from the R package vegan for all comparisons of interest in a data set. Inputs are OTU
and metadata tables, and the output is a .csv. User can tailor contents of .csv by setting perc_cutoff, low_cutoff, and low_val. This function can
also handle taxonomic levels instead of OTU, but currently only select formats are compatible. Requires installation of the R package ‘vegan’.

Inputs:
perc_cutoff: \% cutoff for output OTUs, as decimal (i.e. write 50\% as 0.5), larger \% increases number OTUs in output.
low_cutoff : ‘y’ if want to REMOVE OTUs that contribute less than 1\%
low_val    : set value of low cutoff (0.01), ignored if low_cutoff=‘n’.

Old doc org function:
simper.pretty: automates simper exectution for comparisons of interest
Andrew Steinberger, Kim Dill-Mcfarland, Madison Cox
asteinberger@wisc.edu
Suen Lab
University of Wisconsin-Madison

\if{html}{\out{<div class="sourceCode">}}\preformatted{  Copyright (C) 2016 Andrew Steinberger

This program is free software: you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation, either version 3 of the License, or
(at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with this program.  If not, see <http://www.gnu.org/licenses/>.
}\if{html}{\out{</div>}}

Simper (OTU correlations within inver-metric correlations)

Makes list with the names of all the columns from metrics file that you want to
analyze with SIMPER and combines all results into csv file
}
