% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_resolve_NAs.R
\name{taxa_resolve_NAs}
\alias{taxa_resolve_NAs}
\title{Replace any NAs in taxonomy_table (phyloseq) with known higher level taxonomic data adding tax level k__ p__ f__ g__ etc}
\usage{
taxa_resolve_NAs(ps, verbose = TRUE, genusspecies = TRUE)
}
\arguments{
\item{ps}{phyloseqobject}

\item{verbose}{Default TRUE. Shows diff first 10 taxa before and after}

\item{genusspecies}{Default TRUE. Adds taxonomy column GenusSpecies (Genus+Species) with proper formatting (no prefix, or \link{Genus} etc on different levels avail only)}
}
\value{
NA replaced phyloseq object
}
\description{
Replace any NAs in taxonomy_table (phyloseq) with known higher level taxonomic data adding tax level k__ p__ f__ g__ etc
}
\note{
Version log:
version 1: 20200728
version 2: 20211210 added support species and checks for presence of tax prefix and fixed non-NA issues.
version 3: 20240206 added convenienc taxonomic column GenusSpecies. Fill with Genus+species is species avail (without g__ s__) or with "Genus \link{genus} etc all levels"
}
\author{
a.bossers@uu.nl

alex.bossers@wur.nl
}
