% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toptaxa.R
\name{toptaxa}
\alias{toptaxa}
\title{Returns the top-x list of taxa names at given rank for given samples based on its relative abundances.}
\usage{
toptaxa(ps, rank = "Phylum", top = 10, samples = NA, NArm = FALSE)
}
\arguments{
\item{ps}{phyloseqobject (does not have to be relative abundance, will be converted on the fly)}

\item{rank}{optinal taxonomic rank (default="Phylum").}

\item{top}{optional top number of taxa to return (default 10)}

\item{samples}{optional name of sample or set of sample names to restrict to (default NA, all samples)}
}
\value{
toplist taxa
}
\description{
Returns the top-x list of taxa names at given rank for given samples based on its relative abundances.
}
\author{
Alex.Bossers@wur.nl
Version 1: 20190424'
}
