% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_z_score.R
\name{cal_z_score}
\alias{cal_z_score}
\title{Calculate recalibration matrix for heatmaps
Apply transformation function
Z score is difference to mean divided by sd}
\usage{
cal_z_score(x)
}
\arguments{
\item{x}{value}
}
\value{
tranformed value
}
\description{
Calculate recalibration matrix for heatmaps
Apply transformation function
Z score is difference to mean divided by sd
}
\author{
Alex Bossers \email{a.bossers@uu.nl}
}
