% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdisp.R
\name{hdisp}
\alias{hdisp}
\title{Multiple tests for homogeneity of dispersions (vegan::betadisper)}
\usage{
hdisp(physeq, distm = "bray", vars = NULL, hellinger = TRUE)
}
\arguments{
\item{physeq}{a phyloseq object}

\item{distm}{choice of dissimilarity index according to vegan::vegdist(). Options are "manhattan", "euclidean", "canberra", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup" , "binomial", "chao", "cao" or "mahalanobis".}

\item{vars}{selected variables from the ps object (should be type factor and 2 or more levels). Default uses all.}

\item{hellinger}{default=TRUE. Performs Hellinger transformation on the otu table.}
}
\value{
function prints a list of dataframes for p-values and betadisper output
}
\description{
Function performs individual, betadisper-type tests for homogeneity of dispersions for every element of sample_variables. Community matrix is Hellinger-transformed prior to the calculation of distances using specified matrix.
}
\note{
Alx: added return list incl p-values
}
\examples{
library(phyloseq)
data(GlobalPatterns)
q <- hdisp(GlobalPatterns, distm="bray", vars=c("SampleType"))
plot(q[[1]])

}
\author{
Alex Bossers \email{a.bossers@uu.nl}
}
