% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.permanovas2.R
\name{pairwise.permanovas2}
\alias{pairwise.permanovas2}
\title{Pairwise PERMANOVAs (adonis2) on phyloseq to find which pairs of a more complete set have significantly different community structures}
\usage{
pairwise.permanovas2(
  ps,
  variable,
  sim.function = "vegdist",
  sim.method = "bray",
  p.adjust.m = "BH",
  hellinger = TRUE,
  rngseed = 2202,
  num.cores = "",
  ...
)
}
\arguments{
\item{ps}{phyloseqobject}

\item{sim.function}{function to calculate the similarity matrix (vegdist or daisy). Default is 'vegdist'}

\item{sim.method}{similarity method from daisy or vegdist: default is 'bray'}

\item{p.adjust.m}{p.value correction method, one of the methods supported by p.adjust(); default BH = 'Benjamini Hochberg'}

\item{hellinger}{boolean. Default TRUE to Hellinger transform the "distance" matrix}

\item{rngseed}{The set.seed starting point. Default 2202}

\item{num.cores}{Limit the number of cores to be used. Default = available cores - 1}

\item{metadata}{variable (factor) all factors to be tested pairwise}
}
\value{
a table with the pairwise factors, F-values, R^2, p.value and adjusted p.values
}
\description{
Pairwise PERMANOVAs (adonis2) on phyloseq to find which pairs of a more complete set have significantly different community structures
}
\note{
Based initially on a function by Pedro Martinez Arbizu

Scavenged from https://www.researchgate.net/post/How_can_I_do_PerMANOVA_pairwise_contrasts_in_R
}
\examples{
data(iris)
pairwise.adonis(iris[,1:4],iris$Species)

[1] Signif. codes:  Signif. codes:  0 = ***, 0.001 = **, 0.01 = *, 0.05 = .
                    pairs    F.Model        R2 p.value p.adjusted sig
1    setosa vs versicolor  552.51266 0.8493496   0.001      0.003   *
2     setosa vs virginica 1001.54509 0.9108722   0.001      0.003   *
3 versicolor vs virginica   91.82959 0.4837475   0.001      0.003   *

#similarity euclidean from vegdist and holm correction
#pairwise.adonis(x=iris[,1:4],factors=iris$Species,sim.function='vegdist',sim.method='euclidian',p.adjust.m='holm')
adon=pairwise.adonis2( otu_table(physeq), sample_data(physeq)$Sample_type )

}
\author{
Alex Bossers (a.bossers\@uu.nl)
}
