% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anova_diversity2.R
\name{plot_anova_diversity2}
\alias{plot_anova_diversity2}
\title{Convenience wrapper to ggplot alpha diversity combined with pair-wise ANOVA using a (rarefied) phyloseq object as a starter. Classic or Welch ANOVA supported.}
\usage{
plot_anova_diversity2(
  physeq,
  method,
  rarefy.override = 0,
  grouping_column,
  sample_order = "",
  anova.type = "anova",
  pValueCutoff = 0.05,
  type = "violin",
  whisker = FALSE,
  jitter = TRUE,
  alpha = 0.3,
  x.angle = 0,
  title = "",
  outfile = ""
)
}
\arguments{
\item{physeq}{phyloseq object of microbiome data}

\item{method}{methods can be c("richness", "fisher", "simpson", "invsimpson", "shannon", "evenness", "pd")}

\item{rarefy.override}{Richness needs rarefied data. It automatically detects and rarefies to lowest = default. Override with this setting.}

\item{grouping_column}{Metadata column to use in the grouping}

\item{sample_order}{ordered vector of exact sample/grouping names. Default order of appearance.}

\item{anova.type}{Defaults to classic ANOVA ("anova"). Option to use Welch-ANOVA ("welch").}

\item{pValueCutoff}{default 0.05}

\item{type}{plot type box/violin (default type="violin") else individual points will be shown.}

\item{whisker}{will optionally add whisker boxplot layer to figure. Default = FALSE.}

\item{jitter}{will toggle jitter (random spread 10\% left and 10\% right). Default = TRUE.}

\item{alpha}{plotting dot transparency (default 0.3)}

\item{x.angle}{angle under which x-axis category labels will be printed. Default = 0.}

\item{title}{any title. Default empty.}

\item{outfile}{Optional file name (and path) where to store the data (default "" not saved)}
}
\value{
list of [\link{1}] $plot = ggplot figure, [\link{2}] $anova = pair-wise ANOVA table [\link{3}] $df = underlying data.frame (long format)

Optional tabular data file on disk if \code{outfile} is given
}
\description{
Convenience wrapper to ggplot alpha diversity combined with pair-wise ANOVA using a (rarefied) phyloseq object as a starter. Classic or Welch ANOVA supported.
}
\author{
Alex Bossers \email{a.bossers@uu.nl}

Additional credits go the the original function plot_anova_diversity from \href{https://github.com/umerijaz/microbiomeSeq}{MicrobiomeSeq} by Alfred Ssekagiri; I modified and extended it with some convenience options I found useful to have

Updates:
2023-07-05 alx added x.angle to rotate x-axis legend labels
2022-10-12 alx added Welch-ANOVA options for unequal variances
2022-10-02 alx tweaked box/violin and added \code{whisker} and \code{jitter} options
2022-07-07 alx fixed sample order option to move along significance bars!
2022-07-01 Alx NOT FIXED YET! extended to include ordering of panels (methods)
and samples/grouping avoiding hassling around in ggplot
2022-06-08 Alx extend invSimpson
2019-11-03 Alx Added Type can be violin or box (plot type)
2019-11-02 Alx made error bars more compact using 10\% plot in addition if 10 bars are plotted.
2019-11-01 Alx Also moved over the \code{outfile} option to save table
}
