% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortASVs.R
\name{shortASVs}
\alias{shortASVs}
\title{Move the ASV sequence as sequence_name into the ps refseq slot of the phyloseq object
In addition shorten the taxa names to ASV0001 etc
Default we will also add the ASV seq as lowest tax level in the taxonomy_table}
\usage{
shortASVs(ps, in.taxonomy = TRUE, method = "hash")
}
\arguments{
\item{ps}{phyloseq object (without refseq object and sequences as taxonomy names)}

\item{in.taxonomy}{default TRUE. Adds the actual ASV sequence as lowest level in taxonomy table}

\item{method}{default "hash". Hash will make a hash code of the actual sequence = unique. Alternative ASV\if{html}{\out{<number>}}.}
}
\value{
phyloseq object with ASVs in refseq slot and shortened ASV taxonomy names
}
\description{
Hash method for shortning ASVs is preferred since it would allow still to merge different PS objects since hash 1:1 seq.
By default hash md5 is used. Alternative just a number.
}
\author{
a.bossers@uu.nl
}
