% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.tsv.R
\name{write.tsv}
\alias{write.tsv}
\title{Write data.frame to file using tabular tab separators.}
\usage{
write.tsv(dataToWrite, filename, sep = "\\t", quote = FALSE, row.names = TRUE)
}
\arguments{
\item{dataToWrite}{dataframe}

\item{filename}{filename and path}

\item{sep}{Field separator to use. Default is tab.}

\item{quote}{Boolean default FLASE. Quote fields using ""}

\item{row.names}{Boolean, Add row names. Default TRUE}
}
\description{
Write data.frame to file using tabular tab separators.
}
\author{
Alex Bossers \email{a.bossers@uu.nl}
}
